refactor(ts-migration): migrates ndjsonstream to typescript (#2053)
* feat(ndjsonstream): refactored ndjsonstream and removed old js version

* refactor(submissions.client.factory): updated callsite to use new version of ndjsonstream

* revert(ndjsonstream): adds data_buf back to prevent regression

* fix(ndjsonstream): fixed errorneous initial conditional

* fix(ndjsontream): fixed insidious bug causing occasional off by 1

* revert(ndjsonstream): reverts slice/map to for loops

* revert(ndjsonstream): reverts single line return
